+++ parser_library/src/lsp/completion_list_source.h
@@ -49,7 +49,8 @@ struct completion_list_instructions
 using completion_list_source = std::variant<std::monostate,
     const std::vector<variable_symbol_definition>*,
     const std::unordered_map<context::id_index, std::unique_ptr<context::sequence_symbol>>*,
-    completion_list_instructions>;
+    completion_list_instructions,
+    const context::macro_definition*>;
 
 } // namespace hlasm_plugin::parser_library::lsp