+++ parser_library/src/processing/statement_analyzers/occurrence_collector.h
@@ -39,7 +39,7 @@ public:
 
     occurrence_collector(lsp::occurrence_kind collector_kind,
         context::hlasm_context& hlasm_ctx,
-        lsp::occurrence_storage& storage,
+        std::vector<lsp::symbol_occurrence>& storage,
         bool evaluated_model);
 
     void visit(const semantics::empty_operand& op) override;